+++ Controller/app_main_screen_controller.py
@@ -1,15 +1,15 @@
 """
 controller for main screen
 """
-from configparser import Error
 from typing import List
+from Controller.base_controller import BaseController
 from View.AppMainScreen.app_main_screen_view import AppMainScreenView
 
-from Model.simultaneous_equation_cannons_state import SimultaneousEquationCannonsState
+from Model.simultaneous_equation_cannons_state import CompareMode, SimultaneousEquationCannonsState
 from Model.config_reader import InputConfiguration, read_config, write_config
 
 
-class AppMainScreenController:
+class AppMainScreenController(BaseController):
     """
     The `AppMainScreenController` class represents a controller implementation.
     Coordinates work of the view with the model.
@@ -20,17 +20,12 @@ class AppMainScreenController:
     # technically this is our only data model
     _sec_state = SimultaneousEquationCannonsState([2, 3, 4, 5, 6], [2, 3, 4, 5, 6])
     _config: InputConfiguration = None
+
     def __init__(self, model):
         self.model = model  # Model.app_main_screen.AppMainScreenModel
         self.view = AppMainScreenView(controller=self, model=self.model)
         self.load_simultaneous_equation_cannons_state()
         self.update_view()
-
-    def get_view(self) -> AppMainScreenView:
-        """
-        generated MVC code: controller creates view
-        """
-        return self.view
 
     def load_simultaneous_equation_cannons_state(self):
         """
@@ -40,12 +35,11 @@ class AppMainScreenController:
 
         try:
             self._config = read_config()
-        except Error as e:
+            if self._config is not None:
+                self._sec_state = SimultaneousEquationCannonsState(self._config.fusion_levels, self._config.xyz_ranks)
+        except FileNotFoundError as e:
             print(f"config parser error:\n {e}")
-
-        if self._config is not None:
-            self._sec_state = SimultaneousEquationCannonsState(fusion_levels=sorted(self._config.fusion_levels),
-                                                          xyz_ranks=sorted(self._config.xyz_ranks))
+            self.save_simultaneous_equation_cannons_state()
 
     def save_simultaneous_equation_cannons_state(self):
         """
@@ -55,6 +49,7 @@ class AppMainScreenController:
         input_config.xyz_ranks = self._sec_state.xyz_ranks
         input_config.fusion_levels = self._sec_state.fusion_levels
         write_config(input_config)
+        self._config = input_config
 
     def get_simultaneous_equation_cannons_output(self):
         """
@@ -66,6 +61,8 @@ class AppMainScreenController:
         """
         get config
         """
+        self._config.banished_fusion_levels = self._sec_state.banished_fusion_levels
+        self._config.banished_xyz_ranks = self._sec_state.banished_xyz_ranks
         return self._config
 
     def find_solution(self, monster_level: int, total_cards: int):
@@ -85,6 +82,16 @@ class AppMainScreenController:
         """
         change model save model and update view
         """
-        self._sec_state.set_extra_deck_monster_level(sorted(new_fusion_levels), sorted(new_xyz_ranks))
+        self._sec_state.set_extra_deck_monster_level(new_fusion_levels, new_xyz_ranks)
         self.save_simultaneous_equation_cannons_state()
+        self.update_view()
+
+    def set_banish_zone_monster_level(self, new_fusion_levels: List[int], new_xyz_ranks: List[int]):
+        "add banished xyz fusion monster to the calculation and update view"
+        self._sec_state.set_banish_zone_monster_level(new_fusion_levels, new_xyz_ranks, CompareMode.EXCLUDE)
+        self.update_view()
+
+    def reset_banish_zone_monster_level(self):
+        """remove all banished xyz fusion monsters and update view"""
+        self._sec_state.reset_banish_zone_monster_level()
         self.update_view()