+++ Controller/edit_input_screen_controller.py
@@ -3,10 +3,11 @@ controller for edit input screen,
 edit extra deck monster levels
 """
 from typing import List
-from View.EditInputScreen.edit_input_screen_view import EditInputScreenView
+from Controller.base_controller import BaseController
+from View.EditInputScreen.edit_input_screen_view import EditInputScreenView, InputMode
 
 
-class EditInputScreenController:
+class EditInputScreenController(BaseController):
     """
     The `EditInputScreenController` class represents a controller implementation.
     Coordinates work of the view with the model.
@@ -18,10 +19,6 @@ class EditInputScreenController:
         self.model = model  # Model.edit_input_screen.EditInputScreenModel
         self.view = EditInputScreenView(controller=self, model=self.model)
 
-    def get_view(self) -> EditInputScreenView:
-        """generated get view function for MVC"""
-        return self.view
-
-    def update_extra_deck(self,fusion_levels: List[int], xyz_ranks:List[int]):
+    def update_extra_deck_monsters(self, fusion_levels: List[int], xyz_ranks:List[int], input_mode:InputMode):
         """update extra deck monster level selection in view"""
-        self.view.change_all_levels(fusion_levels,xyz_ranks)
+        self.view.update_extra_deck_monsters(fusion_levels, xyz_ranks, input_mode)