+++ Controller/tutorial_screen_controller.py
@@ -1,8 +1,11 @@
-
+"""
+controller for Tutorial screen
+"""
+from Controller.base_controller import BaseController
 from View.TutorialScreen.tutorial_screen import TutorialScreenView
 
 
-class TutorialScreenController:
+class TutorialScreenController(BaseController):
     """
     The `TutorialScreenController` class represents a controller implementation.
     Coordinates work of the view with the model.
@@ -13,6 +16,3 @@ class TutorialScreenController:
     def __init__(self, model):
         self.model = model  # Model.tutorial_screen.TutorialScreenModel
         self.view = TutorialScreenView(controller=self, model=self.model)
-
-    def get_view(self) -> TutorialScreenView:
-        return self.view