+++ Model/simultaneous_equation_cannons_state.py
@@ -32,19 +32,28 @@ The Fusion Level you send is "the matching Monster level" - the difference of ab
 """
 from typing import List
 from dataclasses import dataclass
-from enum import Enum
+from enum import Enum, unique
 
 
+@unique
 class CardOperation(Enum):
     """add or remove monster level"""
     PLUS = 0
     MINUS = 1
 
 
+@unique
 class MonsterKind(Enum):
     """we only care about fusion and xyz for SimultaneousEquationCannons"""
     FUSION = 0
     XYZ = 1
+
+
+@unique
+class CompareMode(Enum):
+    """if we use banished monster for 2nd SEC calculation, do we exclude xyz monster in banished zone from extra deck"""
+    EXCLUDE = 0
+    NOT_EXCLUDE = 1
 
 
 @dataclass
@@ -63,20 +72,68 @@ class SimultaneousEquationCannonsState():
     _xyz_ranks = [2, 3, 4, 6, 5]
     _value_table = dict()
 
+    _banished_fusion_levels = []
+    _banished_xyz_ranks = []
+    _compare_mode = CompareMode.EXCLUDE
+
     def __init__(self, fusion_levels: List[int], xyz_ranks: List[int]):
         self.set_extra_deck_monster_level(fusion_levels, xyz_ranks)
 
     def _generate_value_table(self):
-        self._value_table = dict()
-        for l in self._fusion_levels:
-            for r in self._xyz_ranks:
-                if l + r in self._value_table:
-                    self._value_table[l + r] = sorted(set(self._value_table[l + r] + [l + r + r]))
-                else:
-                    self._value_table[l + r] = [l + r + r]
+        """
+        generate dictionary with key=monster_level_on_field_to_match and
+        value=total_cards_in_both_players_hand_and_on_both_players_board
 
-    def set_extra_deck_monster_level(self, new_fusion_levels: List[int], new_xyz_ranks: List[int]):
-        """set extra deck input and prepare value table"""
+        monster_level_on_field_to_match can be more than Xyz_Rank + Fusion_Level
+        if there are left over xyz or fusion from past SEC in banished zone
+
+        in that case monster_level_on_field_to_match = every possible combination of 
+        Xyz_Rank and Fusion_Level in Banished Zone
+
+        CompareMode.EXCLUDE means if we have a Rank 4 Xyz Monster Banished, we do not send
+        2 more Rank 4, because we may not have more rank 4 in extra decks.
+        This is the normal case, unless we play more than 2x Rank 4 Monster
+        """
+        self._value_table = dict()
+
+        temp_fusion_levels = self.fusion_levels
+        temp_xyz_ranks = self.xyz_ranks
+        # exclude mode will remove extra deck level/rank if they are also banished
+        if self._compare_mode == CompareMode.EXCLUDE:
+            temp_xyz_ranks = [xyz for xyz in self.xyz_ranks if xyz not in self.banished_xyz_ranks]
+            temp_fusion_levels = [fusion for fusion in self.fusion_levels if fusion not in self.banished_fusion_levels]
+
+        for fusion_lvl in temp_fusion_levels:
+            for xyz_rank in temp_xyz_ranks:
+                total = fusion_lvl + xyz_rank + xyz_rank
+                # add the monster level for empty banished zone
+                self._append_or_add_new_total_in_value_table(fusion_lvl, xyz_rank, total)
+                # add the monster level for banished zone not empty
+                if self._banished_fusion_levels.count or self._banished_xyz_ranks.count:
+                    temp_banished_fusion = self._banished_fusion_levels + [fusion_lvl]
+                    temp_banished_xyz = self._banished_xyz_ranks + [xyz_rank]
+                    for banished_fusion in temp_banished_fusion:
+                        for banished_xyz in temp_banished_xyz:
+                            self._append_or_add_new_total_in_value_table(banished_fusion, banished_xyz, total)
+
+        self._sort_and_remove_duplicate_from_value_table()
+
+    def _sort_and_remove_duplicate_from_value_table(self):
+        # pylint: disable=C0201
+        for k in self._value_table.keys():
+            self._value_table[k] = sorted(set(self._value_table[k]))
+
+    def _append_or_add_new_total_in_value_table(self, fusion_level, xyz_rank, total):
+        """ 
+        append total_card to existing list with key l+r in value_table 
+        or create new list with key l+r in value_table
+        """
+        if fusion_level + xyz_rank in self._value_table:
+            self._value_table[fusion_level + xyz_rank].append(total)
+        else:
+            self._value_table[fusion_level + xyz_rank] = [total]
+
+    def _check_input(self, new_fusion_levels: List[int], new_xyz_ranks: List[int]):
         extra_deck_size = len(new_fusion_levels) + len(new_xyz_ranks) * 2
         if extra_deck_size < 0 or extra_deck_size > 15:
             raise ValueError(f"extra_deck_size out of range: {extra_deck_size}")
@@ -84,12 +141,17 @@ class SimultaneousEquationCannonsState():
         for level in new_fusion_levels:
             if level < 1 or level > 12:
                 raise ValueError(f"fusion level out of range {level}")
-        self._fusion_levels = new_fusion_levels
 
         for rank in new_xyz_ranks:
             if rank < 0 or rank > 13:
                 raise ValueError(f"xyz level out of range {rank}")
-        self._xyz_ranks = new_xyz_ranks
+
+    def set_extra_deck_monster_level(self, new_fusion_levels: List[int], new_xyz_ranks: List[int]):
+        """set extra deck input and prepare value table"""
+        self._check_input(new_fusion_levels, new_xyz_ranks)
+
+        self._fusion_levels = sorted(new_fusion_levels)
+        self._xyz_ranks = sorted(new_xyz_ranks)
         self._generate_value_table()
 
     @property
@@ -101,6 +163,16 @@ class SimultaneousEquationCannonsState():
     def xyz_ranks(self):
         """xyz monster ranks in extra deck"""
         return self._xyz_ranks
+
+    @property
+    def banished_xyz_ranks(self):
+        """banished xyz monster levels"""
+        return self._banished_xyz_ranks
+
+    @property
+    def banished_fusion_levels(self):
+        """banished fusion monster levels"""
+        return self._banished_fusion_levels
 
     @property
     def value_table(self):
@@ -117,8 +189,8 @@ class SimultaneousEquationCannonsState():
     def find_solution(self, monster_level: int, total_cards: int):
         """find out if the banish effect can be activated"""
         res = SimultaneousEquationCannonsSolution(solution_exist=False,
-                                              monster_level=monster_level,
-                                              total_cards=total_cards)
+                                                  monster_level=monster_level,
+                                                  total_cards=total_cards)
         if monster_level not in self._value_table:
             return res
 
@@ -128,7 +200,28 @@ class SimultaneousEquationCannonsState():
         res.solution_exist = True
         res.xyz_rank = total_cards - monster_level
         res.fusion_level = monster_level - res.xyz_rank
-        return res
+
+        temp_xyz_ranks = self.xyz_ranks
+        temp_fusion_levels = self.fusion_levels
+        if self._compare_mode == CompareMode.EXCLUDE:
+            temp_xyz_ranks = [xyz for xyz in self.xyz_ranks if xyz not in self.banished_xyz_ranks]
+            temp_fusion_levels = [fusion for fusion in self.fusion_levels if fusion not in self.banished_fusion_levels]
+        #normal case: if solution without pre banished monster exist
+        if res.xyz_rank in temp_xyz_ranks and res.xyz_rank in temp_fusion_levels:
+            return res
+
+        #special case: if solution only exist with pre banished monsters
+        for xyz in temp_xyz_ranks:
+            for fusion in temp_fusion_levels:
+                if xyz + xyz + fusion == total_cards:
+                    temp_banished_xyz = self.banished_xyz_ranks + [xyz]
+                    temp_banished_fusion = self.banished_fusion_levels + [fusion]
+                    for ban_xyz in temp_banished_xyz:
+                        for ban_fusion in temp_banished_fusion:
+                            if ban_xyz + ban_fusion == monster_level:
+                                res.xyz_rank = xyz
+                                res.fusion_level = fusion
+                                return res
 
     def change_monster(self, level: int, monster_kind: MonsterKind, card_operation: CardOperation):
         """change a single monster in extra deck"""
@@ -151,4 +244,16 @@ class SimultaneousEquationCannonsState():
             return
         self.set_extra_deck_monster_level(temp_fusion_level, temp_xyz_rank)
 
+    def reset_banish_zone_monster_level(self):
+        """remove all banished xyz/fusion monster from calculation"""
+        self._banished_fusion_levels = []
+        self._banished_xyz_ranks = []
+        self._generate_value_table()
 
+    def set_banish_zone_monster_level(self, fusion_levels: List[int], xyz_ranks: List[int], compare_mode: CompareMode):
+        """set banished xyz/fusion monster for calculation"""
+        self._check_input(fusion_levels, xyz_ranks)
+        self._compare_mode = compare_mode
+        self._banished_fusion_levels = sorted(fusion_levels)
+        self._banished_xyz_ranks = sorted(xyz_ranks)
+        self._generate_value_table()