+++ View/AppMainScreen/app_main_screen_view.py
@@ -7,6 +7,7 @@ from kivymd.uix.gridlayout import MDGridLayout
 from kivymd.app import MDApp
 from kivymd.uix.button import MDButton, MDButtonText
 
+from View.EditInputScreen.edit_input_screen_view import InputMode
 from View.base_screen import BaseScreenView
 
 from Model.simultaneous_equation_cannons_state import SimultaneousEquationCannonsSolution
@@ -25,18 +26,20 @@ class CardNumberSelectionButton(MDButton):
         super().__init__(*args, **kwargs)
         self.pos_hint = {"center_x": .5, "center_y": .5}
         self.style = "text"
-        self.md_textfield = (5, 5, 5, 5)
+        #self.md_textfield = (2, 2, 2, 2)
         # self.md_bg_color="#303A29"
         self.radius = 5
-        self.width = 150
+        self.width = 140
         self.theme_width = "Custom"
-
+        #self.size_hint_y = None
+        self.padding= "1dp"
+        self.spacing= "1dp"
     def on_press(self, *args):
         self.selected = not self.selected
         if self.selected:
             self.style = "outlined"
 
-        self.parent.parent.parent.parent.find_solution(self.monster_level, self.total_cards)
+        self.parent.parent.parent.parent.parent.find_solution(self.monster_level, self.total_cards)
 
 
 class AppMainScreenView(BaseScreenView):
@@ -59,7 +62,7 @@ class AppMainScreenView(BaseScreenView):
         if not solution.solution_exist:
             self.ids[['"label_solution"']].text = "No Solution found"
         else:
-            self.ids['"label_solution"'].text = f"xyz rank: {solution.xyz_rank} fusion lvl: {solution.fusion_level }"
+            self.ids['"label_solution"'].text = f"Xyz Rank: {solution.xyz_rank} Fusion lvl: {solution.fusion_level }"
 
     def find_solution(self, monster_level: int, total_cards: int):
         """
@@ -85,9 +88,11 @@ class AppMainScreenView(BaseScreenView):
         """
         self._remove_old_widgets()
         max_col = 0
-        for level in value_table:
+        keys = sorted(value_table.keys())
+
+        for level in keys:
             max_col = max(max_col, len(value_table[level]))
-        for level in value_table:
+        for level in keys:
             # create a section
             grid_layout = MDGridLayout(
                 # pos_hint={"center_x": .5, "top": .88},
@@ -95,22 +100,14 @@ class AppMainScreenView(BaseScreenView):
                 # adaptive_size =  True,
                 # adaptive_height= True,
                 # size_hint_y=  0.05,
-                size_hint_x=1,
-                spacing="1dp",
-                padding="1dp",
+                #size_hint_x=1,
+                spacing="2dp",
+                padding="2dp",
             )
             for total_cards in value_table[level]:
                 card_number_button = CardNumberSelectionButton(
-                    MDButtonText(text=f"level {level} number {total_cards}", text_color="#e6e9df"),
+                    MDButtonText(text=f"Lvl {level} Total {total_cards}", pos_hint={"left": 0,"center_y": .5}),
                     id=f"card_number_button_{level}_{total_cards}",
-                    # theme_width= "Custom",
-                    size_hint_x=0.2,
-                    style="tonal",
-                    md_bg_color="#303A29",
-                    radius=5,
-
-                    width=150,
-                    # height= 25,
                 )
                 card_number_button.monster_level = level
                 card_number_button.total_cards = total_cards
@@ -122,12 +119,12 @@ class AppMainScreenView(BaseScreenView):
         if len(self.to_remove) > 1:
             self.to_remove[0].on_press(None)
 
-    def update_and_go_edit_screen(self):
+    def update_and_go_edit_screen(self, input_mode = InputMode.EXTRA_DECK):
         """
         update edit input screen controller
         and change screen
         """
         app = MDApp.get_running_app()
         config = self.controller.get_simultaneous_equation_cannons_config()
-        app.edit_input_screen_controller.update_extra_deck(config.fusion_levels, config.xyz_ranks)
+        app.edit_input_screen_controller.update_extra_deck(config.fusion_levels, config.xyz_ranks, input_mode)
         app.root.current = "edit_input_screen"