+++ View/constants.py
@@ -1,0 +1,4 @@
+NAME_APP_MAIN_SCREEN = "app_main_screen"
+NAME_EDIT_EXTRA_DECK_SCREEN = "edit_extra_deck_screen"
+NAME_EDIT_BANISHED_ZONE_SCREEN = "edit_banished_zone_screen"
+NAME_TUTORIAL_SCREEN = "tutorial_screen"